Fix DB initialization
The docker-compose config tried to get files needed for DB
initialization from the host (from relative ./my-app/db-config path).
This won't work, because the source files are not copied to the
production server, all assets must be included in the production Docker
image.

New plan: COPY these db-config files to the production docker image in
Github Actions build, and allow access to those files from other
services like "database" and "migrations" using named volumes that map
those release image locations to the named volumes in service
"prod_image" and then again from the named volumes to each of the
reading services' required locations. Bit complex and requires us to
change the dependency structure so that prod_image is actually the first
to run, not the last, but hey it works.