Fix nginx api resolver
Nginx crashes because the postgrest host isn't available yet when nginx
is spun up (postgrest depends on the nginx container, not the other way
around). Using a resolver directive and a little hack with variables
should make nginx wait until there actually is a request to the /api/
endpoint before attempting to resolve the host.

See
https://stackoverflow.com/questions/32845674/setup-nginx-not-to-crash-if-host-in-upstream-is-not-found