Separate nginx from prod container
The database/postgres containers need files that are deployed using the
Dockerfile.prod image, but nginx, which used to be in the
Dockerfile.prod image, needs the postgres container to be on the network
(proxy_pass directive tries to resolve the host eagerly).
This cyclical dependency was pretty annoying to work around (using
variable hacks in nginx conf, see previous commit).

Instead of bundling nginx to our prod image, just use the prod image to
take built html/js files and all the other required configuration to the
server, and then use docker-compose service definitions and named
volumes to make it so that nginx is the last (not the first) service to
be set up, and has access to the files it needs from the prod build
using named volumes.