Clear configuration volumes before starting server
In our production docker-compose, the production Dockerfile invokes COPY
to move deployed files to a named volume, and then other containers will
use these files from that named volume. However, if the volume already
exists in the server, mounting a volume won't work that way and the OLD
files in the volume will be left there instead of getting the new files
from the image.

To get always the NEW bundled files from the image to the volumes, run
docker volume rm on all images except the DB data image (which we'd like
to keep intact) before raising the containers back up with
docker-compose up.